requirements.txt
numpy
fastapi
uvicorn
pydantic
pymongo
python-dotenv
dnspython
python-multipart